ImageHandling/requirements.txt
pytest==7.2.0

Pillow~=9.5.0
future~=0.18.3
PyAutoGUI~=0.9.53